+++ pcs/cli/common/errors.py
@@ -1,12 +1,10 @@
-from typing import (
-    List,
-    Optional,
-)
+from typing import Optional
 
 from pcs.common.str_tools import (
     format_list_base,
     quote_items,
 )
+from pcs.common.types import StringSequence
 
 ERR_NODE_LIST_AND_ALL_MUTUALLY_EXCLUSIVE = (
     "Cannot specify both --all and a list of nodes."
@@ -15,13 +13,17 @@ ERR_NODE_LIST_AND_ALL_MUTUALLY_EXCLUSIVE = (
 SEE_MAN_CHANGES = "See 'man pcs' -> Changes in pcs-{}."
 
 
-def _msg_command_replaced(new_commands: List[str], pcs_version: str) -> str:
+def _msg_command_replaced(
+    new_commands: StringSequence, pcs_version: str
+) -> str:
     commands = format_list_base(quote_items(new_commands))
     changes = SEE_MAN_CHANGES.format(pcs_version)
     return f"This command has been replaced with {commands}. {changes}"
 
 
-def raise_command_replaced(new_commands: List[str], pcs_version: str) -> None:
+def raise_command_replaced(
+    new_commands: StringSequence, pcs_version: str
+) -> None:
     raise CmdLineInputError(
         message=_msg_command_replaced(new_commands, pcs_version=pcs_version)
     )